+++ rosys/__init__.py
@@ -1,8 +1,8 @@
 from .persistence import Persistence
 from .runtime import Runtime
+from . import actors
 from . import communication
 from . import event
+from . import hardware
 from . import pathplanning
 from . import world
-from . import hardware
-from . import actors
