+++ rosys/actors/__init__.py
@@ -2,7 +2,7 @@ from .actor import Actor
 from .automator import Automator
 from .camera_capture import CameraCapture
 from .camera_capture_simulation import CameraCaptureSimulation
+from .detector import Detector
 from .lizard import Lizard
 from .odometer import Odometer
 from .steerer import Steerer
-from .detector import Detector
