+++ rosys/actors/detector.py
@@ -1,8 +1,7 @@
 import socketio
-from ..world import BoxDetection, PointDetection
-from .actor import Actor
 from .. import event, task_logger
-from ..world import Frame
+from ..world import BoxDetection, Frame, PointDetection
+from .actor import Actor
 
 
 class Detector(Actor):
