+++ rosys/world/__init__.py
@@ -1,4 +1,5 @@
 from .camera import Camera, Frame
+from .detection import Detection, BoxDetection, PointDetection
 from .mode import Mode
 from .obstacle import Obstacle
 from .path_segment import PathSegment
@@ -10,4 +11,3 @@ from .rotation import Rotation
 from .spline import Spline
 from .velocity import Velocity
 from .world import World, AutomationState
-from .detection import Detection, BoxDetection, PointDetection
